pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = ""
authors = ["Vadim"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
aiogram = "^3.7.0"
uvicorn = "^0.30.1"
aiohttp = "^3.9.5"
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
pyjwt = "^2.8.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.1"
bcrypt = "^4.1.3"
matplotlib = "^3.9.0"
apscheduler = "^3.10.4"
emoji = "^2.12.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
